rename

+++ processing/interface/file_looper.hh
@@ -59,7 +59,7 @@ private:
     bool _accept_evt(const int& region, const bool& central_unc, const int& jet_cat, const bool& cut_pass, const int& class_id, const float& klambda,
                      const float& cv, const float& c2v, const float& c3);
     double _get_weight(TTreeReaderValue<std::vector<double>>& rv_weight, const std::vector<unsigned int>& idxs);
-    double _get_mva_score(TTreeReaderValue<std::vector<double>>& rv_mva_score, const std::vector<unsigned int>& idxs);
+    float _get_mva_score(TTreeReaderValue<std::vector<float>>& rv_mva_score, const std::vector<unsigned int>& idxs);
 
 public:
     // Methods
